iOS Device: move phyiscal-device methods to a mixin
